% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprintCEE.R
\name{fingerprintCEE}
\alias{fingerprintCEE}
\title{Optimal Fingerprinting with Estimating Equations.}
\usage{
fingerprintCEE(
  Xtilde,
  Y,
  mruns,
  ctlruns.1,
  ctlruns.2,
  nS,
  nT,
  nB = 0,
  conf.level = 0.9,
  cal.a = TRUE,
  missing = FALSE,
  ridge = 0
)
}
\arguments{
\item{Xtilde}{\eqn{n \times p} matrix, signal pattern to be detected.}

\item{Y}{\eqn{n \times 1} matrix, length \eqn{nS \times nT}, observed data.}

\item{mruns}{number of ensembles to estimate the corresponding pattern. 
It is used as the scale of the covariance matrix for \eqn{X_i}.}

\item{ctlruns.1}{\eqn{m \times n} matrix, a group of \eqn{m} independent control 
runs for estimating covariance matrix, which is used in point estimation of 
the signal factors.}

\item{ctlruns.2}{\eqn{m \times n} matrix, another group of \eqn{m} independent control 
runs for estimating the corresponding confidence interval of the signal factors, 
default same as ctlruns.1.}

\item{nS}{number of locations for the observed responses.}

\item{nT}{number of time steps for the observed responses.}

\item{nB}{number of replicates in bootstrap for the bootstrapped variance and confidence 
interval estimations.}

\item{conf.level}{confidence level for confidence interval estimation.}

\item{cal.a}{indicator for calculating the a value, otherwise use default value a = 1.}

\item{missing}{indicator for whether missing values present in Y.}

\item{ridge}{shrinkage value for adjusting the method for missing observations if missing = TRUE.}
}
\value{
a list of the fitted model including point estimate and
interval estimate of beta and a coefficients and corresponding estimate of 
variance, together with the residuals for model diagnostics.
}
\description{
This function estimates the signal factors and corresponding confidence 
interval via the estimating equation methods.
}
\references{
\itemize{ 
\item Sai et al (2023), Optimal Fingerprinting with Estimating Equations,
      \emph{Journal of Climate} 36(20), 7109–-7122.}
}
\author{
Yan Li
}
