% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simDat}
\alias{simDat}
\title{Sample Dataset in Simulation Studies of 
"Regularized Fingerprinting in Detection and Attribution of Climate Change 
with Weight Matrix Optimizing the Efficiency in Scaling Factor Estimation".}
\format{
A data list with two separate data sets.
}
\usage{
data(simDat)
}
\description{
A data list of designed covariance matrix and the expected responses 
to the two forcings ANT and NAT with name
\code{Cov} and \code{X}
where
\itemize{
    \item \code{Cov}: a list of the true covariance matrices;
    \item \code{X}: a data matrix of the expected responses to external forcing;
}
}
\examples{
data(simDat)
}
\keyword{datasets}
